+++ src/components/select.jsx
@@ -1,4 +1,4 @@
-import React, { useCallback, forwardRef } from 'react';
+import React, { useCallback, useMemo, forwardRef } from 'react';
 import PropTypes from 'prop-types';
 import { useNormalisedOptions } from '../hooks/use_normalised_options';
 import { renderGroupedOptions } from '../helpers/render_grouped_options';
@@ -10,29 +10,44 @@ export const Select = forwardRef((rawProps, ref) => {
   });
 
   const {
+    multiple,
     onChange,
     onValue,
+    onValues,
     options,
     renderOption,
     renderOptGroup,
     selectedOption,
+    selectedOptions,
     value: _0,
-    placeholderOption: _1,
-    mapOption: _2,
-    nullOptions: _3,
+    values: _1,
+    placeholderOption: _2,
+    mapOption: _3,
+    nullOptions: _4,
     ...selectProps
   } = optionisedProps;
 
   const handleChange = useCallback((e) => {
+    onValues?.([...e.target.selectedOptions].map((el) => (
+      options.find((o) => o.identity === el.value)?.value
+    )));
     onValue?.(options.find((o) => o.identity === e.target.value)?.value);
     onChange?.(e);
-  }, [onValue, onChange, options]);
+  }, [onValue, onValues, onChange, options]);
+
+  const reactValue = useMemo(() => {
+    if (multiple) {
+      return selectedOptions.map((o) => o?.identity);
+    }
+    return selectedOption?.identity ?? '';
+  }, [selectedOption, selectedOptions, multiple]);
 
   return (
     <select
-      value={selectedOption?.identity ?? ''}
+      value={reactValue}
       onChange={handleChange}
       ref={ref}
+      multiple={multiple}
       {...selectProps}
     >
       {renderGroupedOptions({
@@ -64,19 +79,25 @@ export const Select = forwardRef((rawProps, ref) => {
 
 Select.propTypes = {
   placeholderOption: PropTypes.string,
+  multiple: PropTypes.bool,
   onChange: PropTypes.func,
   onValue: PropTypes.func,
+  onValues: PropTypes.func,
   options: PropTypes.arrayOf(PropTypes.any).isRequired,
   value: PropTypes.any,
+  values: PropTypes.arrayOf(PropTypes.any),
   renderOption: PropTypes.func,
   renderOptGroup: PropTypes.func,
 };
 
 Select.defaultProps = {
   placeholderOption: null,
+  multiple: false,
   value: null,
+  values: null,
   onChange: null,
   onValue: null,
+  onValues: null,
   renderOption: (props) => <option {...props} />,
   renderOptGroup: (props) => <optgroup {...props} />,
 };